Fix statusHistory
- Moved statusHistory to cli/history.go
- Flush tabwriter
